Fallback to crate_version when not in git repository
When not building in a git repository, `git_version::git_version!()` would not
compile, because it requires to be run in a git repository to get the
current tag/branch/hash/status. [1]

This adds a fallback to the macro by using `clap::crate_version!()` which gets the
version from `Cargo.toml`.

[1] One of these situations is when running `cargo install dotfilers`
instead of cloning the repository and running `cargo build`.